:white_check_mark: Add simple test for Post component.
